テスト構成を一本化
